requirements.txt

# requirements cahtdoc
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
llama-index==0.9.34
llama_cpp_python==0.1.83
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.9.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==4.0.0
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
transformers==4.37.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
wrapt==1.16.0
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
# requirements scarpxiv
xmltodict==0.12.0
requests==2.31.0
pandas==2.2.0
pdftotext==2.2.2
jupyter
ipykernel==6.29.3